Unescape dates with i18n to render them in a readable format
One can create a custom interpolation function for Date instances. [[1]]

However, if those dates include characters like "/", the returned string
will be escaped. This is the default behavior for string interpolation
in i18next. [[2]]

Since we are only rendering a Date, and it does not present an
opportunity for a malicious actor to perform a XSS attack, we can
unescape this value to render the date in a readable format (ex.
08/10/2021).

[1]: https://www.i18next.com/translation-function/formatting
[2]: https://www.i18next.com/translation-function/interpolation#unescape